fix: extract fleet_run to _fleet_run.py to satisfy line-limit + fix ruff E501
fleet/__init__.py exceeded the 415-line arch gate after fleet_run was added.
Extracted fleet_run and _fleet_run_error to _fleet_run.py; __init__.py now
registers the command via fleet_app.command(name="run")(fleet_run).
Also shortened three docstrings in test_fleet_run.py that exceeded 99 chars.
