+++ src/autoskillit/cli/fleet/__init__.py
@@ -9,7 +9,7 @@ import shutil
 import sys
 from datetime import UTC, datetime
 from pathlib import Path
-from typing import TYPE_CHECKING, Annotated, NoReturn
+from typing import TYPE_CHECKING, Annotated
 from uuid import uuid4
 
 from cyclopts import App, Parameter
@@ -31,6 +31,7 @@ from autoskillit.cli.fleet._fleet_preview import (
     _FLEET_DISPATCH_GREETINGS,
     _print_dispatch_preview,
 )
+from autoskillit.cli.fleet._fleet_run import fleet_run
 from autoskillit.cli.fleet._fleet_session import _launch_fleet_session
 from autoskillit.core import TerminalColumn, get_logger, is_feature_enabled
 
@@ -49,21 +50,6 @@ def _require_fleet(cfg: AutomationConfig) -> None:
             file=sys.stderr,
         )
         raise SystemExit(1)
-
-
-def _fleet_run_error(error: str, message: str, exit_code: int = 1) -> NoReturn:
-    """Print a JSON error envelope to stdout and exit nonzero.
-
-    Headless-only error helper: emits machine-readable JSON so dispatch
-    callers can parse the refusal rather than scraping stderr text.
-    """
-    envelope = {
-        "success": False,
-        "error": error,
-        "user_visible_message": message,
-    }
-    print(json.dumps(envelope))
-    raise SystemExit(exit_code)
 
 
 if TYPE_CHECKING:
@@ -425,67 +411,4 @@ def fleet_status(
         print(_render_terminal_table(columns, rows_list))
 
 
-@fleet_app.command(name="run")
-def fleet_run(
-    recipe: str,
-    *,
-    task: Annotated[str, Parameter(name=["--task", "-t"])] = "",
-    ingredient: Annotated[tuple[str, ...], Parameter(name=["--ingredient", "-i"])] = (),
-    backend: Annotated[str | None, Parameter(name=["--backend"])] = None,
-    timeout_sec: Annotated[int | None, Parameter(name=["--timeout-sec"])] = None,
-    resume_session_id: Annotated[str | None, Parameter(name=["--resume-session-id"])] = None,
-    prior_dispatch_id: Annotated[str | None, Parameter(name=["--prior-dispatch-id"])] = None,
-    disable_quota_guard: Annotated[bool, Parameter(name=["--disable-quota-guard"])] = False,
-) -> None:
-    """One-shot headless recipe dispatch (experimental).
-
-    Dispatches a single recipe run non-interactively. Prints the dispatch
-    result envelope as JSON on stdout. Exit 0 on SUCCESS, nonzero otherwise.
-    """
-    # --- Session-type guard (retained for headless — prevents recursive dispatch) ---
-    if os.environ.get("AUTOSKILLIT_SESSION_TYPE") in ("skill", "leaf"):
-        _fleet_run_error(
-            "FLEET_SESSION_TYPE_BLOCKED",
-            "'fleet run' cannot run inside a skill or leaf session.",
-        )
-    # NOTE: CLAUDECODE guard intentionally NOT applied — REQ-HFD-006
-
-    # --- Config + feature gates ---
-    from autoskillit.config import load_config
-
-    try:
-        cfg = load_config(Path.cwd())
-    except Exception as exc:
-        logger.error("fleet run: failed to load config", exc_info=True)
-        _fleet_run_error("FLEET_CONFIG_ERROR", f"Failed to load config: {exc}")
-
-    # Fleet base feature check (equivalent to _require_fleet but with JSON output)
-    if not is_feature_enabled(
-        "fleet", cfg.features, experimental_enabled=cfg.experimental_enabled
-    ):
-        _fleet_run_error(
-            "FLEET_FEATURE_DISABLED",
-            "The 'fleet' feature is not enabled.\n"
-            "Enable with: features.experimental_enabled: true in your config\n"
-            "Or set: AUTOSKILLIT_FEATURES__FLEET=true",
-        )
-
-    # Headless run feature check
-    if not is_feature_enabled(
-        "fleet_headless_run",
-        cfg.features,
-        experimental_enabled=cfg.experimental_enabled,
-    ):
-        _fleet_run_error(
-            "FLEET_FEATURE_DISABLED",
-            "The 'fleet_headless_run' feature is not enabled.\n"
-            "Enable with: features.experimental_enabled: true\n"
-            "Or: features.fleet_headless_run: true\n"
-            "Or: AUTOSKILLIT_FEATURES__FLEET_HEADLESS_RUN=true",
-        )
-
-    # --- Part B implements the dispatch body here ---
-    _fleet_run_error(
-        "FLEET_NOT_IMPLEMENTED",
-        "'fleet run' dispatch is not yet implemented. See Part B.",
-    )
+fleet_run = fleet_app.command(name="run")(fleet_run)
